deployment/requirements.txt
streamlit == 1.30.0
pandas == 1.4.4
seaborn == 0.11.2
matplotlib == 3.5.2
plotly == 5.9.0
numpy == 1.21.5
Pillow == 9.2.0
scikit.learn == 1.0.2